requirements.txt
requests
python-dotenv
colored
psycopg2-binary